Version v0.16.1 - CHANGELOG.md [citest skip]
[v0.16.1] - 2022-07-28
--------------------

- add parameter RSAMinSize

Add support for the new RSAMinSize parameter.

- Ensure values are cast to correct type

- Addition notes about secondary variables
- Fix various linting issues
- Revert incorrect module name
- tests: Do not be picky about spaces/tabs

When testing with cloud-init, it modifies the sshd_configuration and can
replace some tabs with whitespaces. This happens frequently around the
subsystem keyword. There are no functional changes, but the matching
did not work as expected.

Signed-off-by: Jakub Jelen <jjelen@redhat.com>

- the role still supports ansible 2.9

- Add CHANGELOG.md

- Add changelog_to_tag.yml to .github/workflows

Description:
When a new changelog section is added to CHANGELOG.md and pushed,
changelog_to_tag.yml is triggered, which generates a new tag and
a new release.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>