fix: Document and streamline the sshd_main_config_file
The option was introduced in 6bb0d7b456eececcf5b5e0c817b113cfde98d170
without documentation and intended use. The recent change
f6ae2094fe23eed94ce64eee1d0f5f1109dfe1da propagated this option to the
generated service files, which is resulting in unexpected results, when
a user decided to set only `sshd_config_file` for the second sshd
service causing the service file points to the system-wide
configuration file.

This is an attempt to fix this by introducing some heuristics to guess
if the user wants to set up second drop-in directory (ending with .d)
or create a standalone configuration file.

Fixes: #280