ci: use ansible-lint v25 - have to use requirements file for ansible-lint now (#309)
This PR supersedes https://github.com/willshersystems/ansible-sshd/pull/306
because there is some extra work that must be done to use v25
We need to provide a requirements file.  Even though sshd does not have
a tests collection-requirements.yml file, we use the same code we
use in linux-system-roles that handles/merges the meta and tests
requirements files.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>